+++ libfreeipmi/include/freeipmi/cmds/ipmi-fru-inventory-device-cmds.h
@@ -16,7 +16,7 @@
   Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA.
 */
 /*****************************************************************************\
-*  $Id: ipmi-fru-inventory-device-cmds.h,v 1.7 2009-04-17 23:50:26 chu11 Exp $
+*  $Id: ipmi-fru-inventory-device-cmds.h,v 1.8 2009-04-29 00:05:47 chu11 Exp $
 *****************************************************************************
 *  Copyright (C) 2007-2009 Lawrence Livermore National Security, LLC.
 *  Copyright (C) 2007 The Regents of the University of California.
@@ -69,19 +69,19 @@ extern fiid_template_t tmpl_cmd_read_fru_data_rs;
 extern fiid_template_t tmpl_cmd_write_fru_data_rq;
 extern fiid_template_t tmpl_cmd_write_fru_data_rs;
 
-int8_t fill_cmd_get_fru_inventory_area_info (uint8_t fru_device_id,
-                                             fiid_obj_t obj_cmd_rq);
+int fill_cmd_get_fru_inventory_area_info (uint8_t fru_device_id,
+                                          fiid_obj_t obj_cmd_rq);
 
-int8_t fill_cmd_read_fru_data (uint8_t fru_device_id,
-                               uint16_t fru_inventory_offset_to_read,
-                               uint8_t count_to_read,
-                               fiid_obj_t obj_cmd_rq);
+int fill_cmd_read_fru_data (uint8_t fru_device_id,
+                            uint16_t fru_inventory_offset_to_read,
+                            uint8_t count_to_read,
+                            fiid_obj_t obj_cmd_rq);
 
-int8_t fill_cmd_write_fru_data (uint8_t fru_device_id,
-                                uint16_t fru_inventory_offset_to_write,
-                                uint8_t *data_to_write,
-                                unsigned int data_to_write_len,
-                                fiid_obj_t obj_cmd_rq);
+int fill_cmd_write_fru_data (uint8_t fru_device_id,
+                             uint16_t fru_inventory_offset_to_write,
+                             uint8_t *data_to_write,
+                             unsigned int data_to_write_len,
+                             fiid_obj_t obj_cmd_rq);
 
 #ifdef __cplusplus
 }
